#2.10

+++ api/questions_api.py
@@ -4,13 +4,22 @@ from data.categories import Category
 from data import db_session
 import os
 from secondary_functions import get_time
+import configparser
+
+'''Cоздаём объекта парсера. Читаем конфигурационный файл'''
+config = configparser.ConfigParser()
+config.read("config.ini", encoding='utf-8')
+
+"""Загрузка переменных среды из файла"""
+from dotenv import load_dotenv
+load_dotenv(dotenv_path=config['PATH']['SECRET_DATA'])
 
 
 blueprint = Blueprint('questions_api', __name__, template_folder='templates')
 
 
 '''API для получения одного вопроса'''
-@blueprint.route('/api/123456789/question/<int:question_id>',  methods=['GET'])
+@blueprint.route(f'/api/{ os.getenv("TOKEN") }/question/<int:question_id>',  methods=['GET'])
 def get_one_question(question_id):
     session = db_session.create_session()
     questions = session.query(Question).get(question_id)
@@ -25,7 +34,7 @@ def get_one_question(question_id):
 
 
 '''API для получения всех вопросов'''
-@blueprint.route('/api/123456789/questions',  methods=['GET'])
+@blueprint.route(f'/api/{ os.getenv("TOKEN") }/questions',  methods=['GET'])
 def get_questions():
     session = db_session.create_session()
     questions = session.query(Question).all()
@@ -49,30 +58,58 @@ def get_questions():
 
 
 '''API для создания вопроса'''
-@blueprint.route('/api/123456789/add_quest', methods=['POST'])
+@blueprint.route(f'/api/{ os.getenv("TOKEN") }/add_quest', methods=['POST'])
 def create_questions():
 
-    if not request.json:
-        return jsonify({'error': 'Empty request'})
-    elif not all(key in request.json for key in
-                 ['id', 'text', 'category', 'answers', 'right_answer', 'who_add']):
-        return jsonify({'error': 'Bad request'})
     session = db_session.create_session()
-    quest = Question(
-        id=request.json['id'],
-        text=request.json['text'],
-        category=request.json['category'],
-        answers=request.json['answers'],
-        right_answer=request.json['right_answer'],
-        who_add=request.json['who_add']
-    )
-    session.add(quest)
-    session.commit()
+
+    try:
+        print(request.form)
+        quest = session.query(Question).filter(Question.text == request.form['text']).first()  # 7
+        if quest:
+            return jsonify({'errors': 'Такой вопрос уже есть в базе'})
+
+        quest = Question(
+            text=request.form['text'],
+            comment=request.form['comment'],
+            right_answer=request.form['answer'],
+            answers='!@#$%'.join([request.form['answer'].split('!@#$%')[0],
+                     request.form['wrong_answer1'],
+                     request.form['wrong_answer2'],
+                     request.form['wrong_answer3']]),
+            who_add=request.form['who_add'],
+            category=session.query(Category).filter(Category.name == request.form['category']).first().id,
+            type=request.form['type'],
+            complexity=request.form['complexity'],
+            is_promoted=1 if request.form['state_who_add'] == 'admin' else 0)
+
+        session.add(quest)
+        session.commit()
+
+        file = request.files['file'].read()
+        if str(file) != "b''":
+            quest.images = f'/static/img/questions/{quest.id}.png'
+            with open(quest.images[1:], 'wb') as f1:  # 10
+                f1.write(file)
+        else:
+            quest.images = ' '
+
+        session.commit()
+    except Exception as e:
+        print(e)
+        return jsonify({'errors': 'Неизвестная ошибка'})
+
+    try:
+        session.add(quest)
+        session.commit()
+    except Exception as e:
+        print(e)
+        return jsonify({'errors': 'Неизвестная ошибка'})
     return jsonify({'success': 'OK'})
 
 
 '''API для удаления вопроса'''
-@blueprint.route('/api/123456789/quests/<int:quest_id>', methods=['DELETE'])
+@blueprint.route(f'/api/{ os.getenv("TOKEN") }/quests/<int:quest_id>', methods=['DELETE'])
 def delete_questions(quest_id):
     session = db_session.create_session()
     quest = session.query(Question).get(quest_id)
@@ -84,9 +121,10 @@ def delete_questions(quest_id):
 
 
 '''API для редактирования вопроса'''
-@blueprint.route('/api/123456789/quests/<int:quest_id>', methods=['PUT'])
+@blueprint.route(f'/api/{ os.getenv("TOKEN") }/quests/<int:quest_id>', methods=['PUT'])
 def put_questions(quest_id):
     session = db_session.create_session()
+    print(1)
     quest = session.query(Question).get(quest_id)
     if not quest:
         return jsonify({'error': 'Not found'})
@@ -106,7 +144,7 @@ def put_questions(quest_id):
                 and request.form.get('wrong_answer2_' + str(quest_id))\
                 and request.form.get('wrong_answer3_' + str(quest_id)):
             print(4)
-            quest.answers = '!@#$%'.join([request.form['answer_' + str(quest_id)].strip(),
+            quest.answers = '!@#$%'.join([request.form['answer_' + str(quest_id)].strip().split('!@#$%')[0],
                                           request.form['wrong_answer1_' + str(quest_id)].strip(),
                                           request.form['wrong_answer2_' + str(quest_id)].strip(),
                                           request.form['wrong_answer3_' + str(quest_id)].strip()])
@@ -120,11 +158,17 @@ def put_questions(quest_id):
             print(7)
             quest.complexity = request.form['select_comp_question_edit_redactor_' + str(quest_id)]
         if request.files.get('image_' + str(quest_id)):
-            print(8)
-            os.remove(quest.images[1:])
-            quest.images = f'/static/img/questions/{quest.id}+{get_time()}.png'
-            with open(quest.images[1:], 'wb') as f1:  # 10
-                f1.write(request.files['image_' + str(quest_id)].read())
+            file = request.files['image_' + str(quest_id)].read()
+            if str(file) != "b''":
+                if quest.images[1:]:
+                    os.remove(quest.images[1:])
+                quest.images = f'/static/img/questions/{quest.id}+{get_time()}.png'
+                with open(quest.images[1:], 'wb') as f1:  # 10
+                    f1.write(file)
+            elif request.form.get('hidden_image_' + str(quest_id)):
+                if request.form['hidden_image_' + str(quest_id)] == 'no_image':
+                    quest.images = ' '
+
         session.commit()
     except Exception as e:
         print(e)
